Color detection sample with opencv and numpy
This project uses the opencv and numpy libraries to detect colors and is based on the project: REAL TIME COLOR DETECTION With Python and OpenCV | Detection of Fruits by Color in Python (https://youtu.be/YTEu7ybnHgE) from the Aprende e Ingenia channel (https://www.youtube.com/@AprendeIngenia) (Spanish)

It is possible to manually determine the range of HSV values of any object that you want to detect. It uses a kernel to perform a morphological transformation to smooth the image. I hope this little project is useful.